+++ rulesets/tests/test_es_enclosure.rb
@@ -98,7 +98,7 @@ class EnergyStarEnclosureTest < MiniTest::Test
     ESConstants.NationalVersions.each do |es_version|
       _convert_to_es('base.xml', es_version)
       hpxml = HPXML.new(hpxml_path: @tmp_hpxml_path)
-      hpxml.climate_and_risk_zones.iecc_zone = '1A'
+      hpxml.climate_and_risk_zones.climate_zone_ieccs[0].zone = '1A'
       hpxml.climate_and_risk_zones.weather_station_name = 'Miami, FL'
       hpxml.climate_and_risk_zones.weather_station_wmo = 722020
       XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
@@ -248,11 +248,11 @@ class EnergyStarEnclosureTest < MiniTest::Test
       end
 
       hpxml_names = ['base.xml',
-                     'base-foundation-conditioned-basement-wall-interior-insulation.xml']
+                     'base-foundation-conditioned-basement-wall-insulation.xml']
       hpxml_names.each do |hpxml_name|
         _convert_to_es(hpxml_name, es_version)
         hpxml = _test_ruleset()
-        if hpxml_name == 'base-foundation-conditioned-basement-wall-interior-insulation.xml'
+        if hpxml_name == 'base-foundation-conditioned-basement-wall-insulation.xml'
           type = HPXML::FoundationWallTypeConcreteBlockFoamCore
         else
           type = nil
@@ -269,16 +269,16 @@ class EnergyStarEnclosureTest < MiniTest::Test
       end
 
       hpxml_names = ['base.xml',
-                     'base-foundation-conditioned-basement-wall-interior-insulation.xml']
+                     'base-foundation-conditioned-basement-wall-insulation.xml']
       hpxml_names.each do |hpxml_name|
         _convert_to_es(hpxml_name, es_version)
         hpxml = HPXML.new(hpxml_path: @tmp_hpxml_path)
-        hpxml.climate_and_risk_zones.iecc_zone = '1A'
+        hpxml.climate_and_risk_zones.climate_zone_ieccs[0].zone = '1A'
         hpxml.climate_and_risk_zones.weather_station_name = 'Miami, FL'
         hpxml.climate_and_risk_zones.weather_station_wmo = 722020
         XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
         hpxml = _test_ruleset()
-        if hpxml_name == 'base-foundation-conditioned-basement-wall-interior-insulation.xml'
+        if hpxml_name == 'base-foundation-conditioned-basement-wall-insulation.xml'
           type = HPXML::FoundationWallTypeConcreteBlockFoamCore
         else
           type = nil
@@ -397,7 +397,7 @@ class EnergyStarEnclosureTest < MiniTest::Test
 
       _convert_to_es('base-foundation-unconditioned-basement.xml', es_version)
       hpxml = HPXML.new(hpxml_path: @tmp_hpxml_path)
-      hpxml.climate_and_risk_zones.iecc_zone = '1A'
+      hpxml.climate_and_risk_zones.climate_zone_ieccs[0].zone = '1A'
       hpxml.climate_and_risk_zones.weather_station_name = 'Miami, FL'
       hpxml.climate_and_risk_zones.weather_station_wmo = 722020
       XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
@@ -447,7 +447,7 @@ class EnergyStarEnclosureTest < MiniTest::Test
 
       _convert_to_es('base-foundation-slab.xml', es_version)
       hpxml = HPXML.new(hpxml_path: @tmp_hpxml_path)
-      hpxml.climate_and_risk_zones.iecc_zone = '1A'
+      hpxml.climate_and_risk_zones.climate_zone_ieccs[0].zone = '1A'
       hpxml.climate_and_risk_zones.weather_station_name = 'Miami, FL'
       hpxml.climate_and_risk_zones.weather_station_wmo = 722020
       XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
@@ -582,7 +582,7 @@ class EnergyStarEnclosureTest < MiniTest::Test
 
       _convert_to_es('base.xml', es_version)
       hpxml = HPXML.new(hpxml_path: @tmp_hpxml_path)
-      hpxml.climate_and_risk_zones.iecc_zone = '1A'
+      hpxml.climate_and_risk_zones.climate_zone_ieccs[0].zone = '1A'
       hpxml.climate_and_risk_zones.weather_station_name = 'Miami, FL'
       hpxml.climate_and_risk_zones.weather_station_wmo = 722020
       XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)