+++ rulesets/tests/test_ventilation.rb
@@ -920,7 +920,7 @@ class ERIMechVentTest < MiniTest::Test
   def test_mech_vent_iecc_eri_exception
     IECCConstants.AllVersions.each do |iecc_version|
       _all_calc_types.each do |calc_type|
-        hpxml_name = 'base-mechvent-exhaust.xml'
+        hpxml_name = _change_iecc_version('base-mechvent-exhaust.xml', iecc_version)
 
         # Run IECC calculation
         iecc_hpxml = _test_ruleset(hpxml_name, calc_type, iecc_version)
@@ -931,7 +931,8 @@ class ERIMechVentTest < MiniTest::Test
 
         if ['2018', '2021'].include?(iecc_version) && calc_type == Constants.CalcTypeERIReferenceHome
           # Check that ventilation exception in 2018/2021 IECC is being applied to the ERI Reference Home
-          assert_operator(iecc_hpxml.ventilation_fans[0].tested_flow_rate, :<, base_hpxml.ventilation_fans[0].tested_flow_rate)
+          refute_equal(iecc_hpxml.ventilation_fans[0].tested_flow_rate, base_hpxml.ventilation_fans[0].tested_flow_rate)
+          assert_equal(57.0, iecc_hpxml.ventilation_fans[0].tested_flow_rate)
         else
           # In all other cases, check for the same ventilation as the standard ERI
           assert_equal(base_hpxml.ventilation_fans[0].tested_flow_rate, iecc_hpxml.ventilation_fans[0].tested_flow_rate)